Added save/load and EXP calculation!
There's still styling and UX stuff that should be done, but barring breaking bugs, this SHOULD be kind of a decent proof of concept for the tool.

Known bug is that entries aren't sorted and EXP isn't calculated on load; but that's not likely to be a huge issue for most users since they won't have data without it already there.